% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour3d.R
\name{contour3d}
\alias{contour3d}
\title{Compute Isosurface, a Three Dimension Contour}
\usage{
contour3d(griddata, level, x, y, z)
}
\arguments{
\item{griddata}{A three dimensional array from which to calculate the contour}

\item{level}{The level at which to construct the contour surface}

\item{x, y, z}{locations of grid planes at which values in \code{griddata} are measured}
}
\value{
Returns a list with coordinates of each surface vertex, indices of
the vertices that make up each triangle, and surface normals at each vertex
}
\description{
Computes a 3D contours or isosurface by the marching cubes algorithm.
}
