\name{rmarchingcubes-package}
\alias{rmarchingcubes-package}
\alias{rmarchingcubes}
\docType{package}
\title{
\packageTitle{rmarchingcubes}
}
\description{
\packageDescription{rmarchingcubes}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{rmarchingcubes}
\packageIndices{rmarchingcubes}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{rmarchingcubes}

Maintainer: \packageMaintainer{rmarchingcubes}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
